pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-bot"
version = "0.1.0"
description = "Crypto trading bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "PyYAML>=6.0",
  "python-binance>=1.0.19",
  "python-dotenv>=1.0.0",
  "ta>=0.11.0",
  "Flask>=3.0",
]

[tool.setuptools]
packages = []
